Allow Snapping to have an empty return
I created a function without a return value. this lead to the error of "x" being undefined and the parent not being recognized.
And since all other functions don't need a return value, I thought one may handle the empty function within the repository.

In my opinion the snap should be explicitly disallowed by returning false. Most likely this can be seen as a design choice.
So it may not be to your liking. Since the changes are minimal I thought I would go ahead and make a PR.